# docs/openapi.yaml
openapi: 3.0.3
info:
  title: POMS API
  description: |
    **Print Operations Management System** — REST API for managing print job orders,
    staff tasks, inventory, billing, payments, and analytics.

    ## Authentication
    Every endpoint except `POST /auth/login` requires a Bearer JWT in the
    `Authorization` header.

    ## Error shape
    All errors return `{ "error": "<message>" }` with the appropriate HTTP status.

    ## Roles & permissions
    | Role | orders | tasks | billing | payments | inventory | machines | analytics | users |
    |------|--------|-------|---------|----------|-----------|----------|-----------|-------|
    | super_admin | CRUD | CRUD | CRUD | CRUD | CRUD | CRUD | R | CRUD |
    | admin | CRU | CRUD | R | CRU | CRU | CRU | R | CRU |
    | sub_admin | CRU | CRU | R | R | R | R | — | R |
    | designer | R | CRU (own) | — | — | R | — | — | — |
    | *_staff | R (own) | RU (own) | — | — | R | — | — | — |

  version: 1.0.0
  contact:
    email: capitalcurv8@gmail.com
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable parameters ────────────────────────────────────────────────────
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: 6655a1b2c3d4e5f6a7b8c9d0
    orderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        example: 6655a1b2c3d4e5f6a7b8c9d0

  # ─── Reusable responses ──────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
    BadRequest:
      description: Validation error or invalid state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Duplicate unique field
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ─── Schemas ────────────────────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth ────────────────────────────────────────────────────────────────────
...
    Role:
      type: string
      enum:
        - super_admin
        - admin
        - sub_admin
        - designer
        - flex_printing_staff
        - screen_printing_staff
        - helper_staff
        - expose_staff
        - binder_staff
        - offset_staff
        - cutting_staff
        - laser_cut_staff
        - acrylic_printing_staff
        - glass_printing_staff
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        phone:
          type: string
        isAvailable:
          type: boolean
        activeTaskCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/Role'
        phone:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        phone:
          type: string
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer

    # Orders ──────────────────────────────────────────────────────────────────
    JobType:
      type: string
      enum:
        - flex_printing
        - screen_printing
        - design
        - laser_cut
        - offset
        - acrylic
        - glass
        - binding
    OrderStatus:
      type: string
      enum:
        - draft
        - confirmed
        - designing
        - in_production
        - finishing
        - completed
        - invoiced
        - cancelled
    Priority:
      type: string
      enum: [normal, urgent]
    OrderItem:
      type: object
      required: [description, quantity, unit, unitPrice]
      properties:
        description:
          type: string
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          example: sqft
        unitPrice:
          type: number
          minimum: 0
    BomEntry:
      type: object
      properties:
        materialId:
          type: string
        name:
          type: string
        unit:
          type: string
        qty:
          type: number
    StatusHistory:
      type: object
      properties:
        status:
          type: string
        changedBy:
          type: string
        note:
          type: string
        at:
          type: string
          format: date-time
    Order:
      type: object
      properties:
        _id:
          type: string
        orderNumber:
          type: string
          example: ORD-2024-0001
        customer:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            email:
              type: string
              format: email
        jobType:
          $ref: '#/components/schemas/JobType'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        bom:
          type: array
          items:
            $ref: '#/components/schemas/BomEntry'
        rawCost:
          type: number
        taxableValue:
          type: number
        billSplitPct:
          type: number
          minimum: 0
          maximum: 100
        hsnCode:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/StatusHistory'
        priority:
          $ref: '#/components/schemas/Priority'
        deadline:
          type: string
          format: date-time
        advancePaid:
          type: number
        notes:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [customer, jobType, items, rawCost, taxableValue, billSplitPct]
      properties:
        customer:
          type: object
          required: [name, phone]
          properties:
            name:
              type: string
              minLength: 2
            phone:
              type: string
              minLength: 7
            email:
              type: string
              format: email
        jobType:
          $ref: '#/components/schemas/JobType'
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        bom:
          type: array
          items:
            $ref: '#/components/schemas/BomEntry'
        rawCost:
          type: number
          minimum: 0
        taxableValue:
          type: number
          minimum: 0
        billSplitPct:
          type: number
          minimum: 0
          maximum: 100
        hsnCode:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        deadline:
          type: string
          format: date-time
        notes:
          type: string
    UpdateOrderRequest:
      type: object
      properties:
        customer:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            email:
              type: string
              format: email
        rawCost:
          type: number
        taxableValue:
          type: number
        billSplitPct:
          type: number
        hsnCode:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        deadline:
          type: string
          format: date-time
        notes:
          type: string
    SetOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
        note:
          type: string
    OrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total:
          type: integer

    # Tasks ───────────────────────────────────────────────────────────────────
...
    PaymentType:
      type: string
      enum: [advance, final, partial]
    PaymentMethod:
      type: string
      enum: [cash, upi, card]
    PaymentStatus:
      type: string
      enum: [pending, completed, refunded]
    Payment:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
        billId:
          type: string
        type:
          $ref: '#/components/schemas/PaymentType'
        amount:
          type: number
        method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        referenceId:
          type: string
        collectedBy:
          type: string
        paidAt:
          type: string
          format: date-time
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...
    PaymentList:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        total:
          type: integer

    # Billing ─────────────────────────────────────────────────────────────────
    BillType:
      type: string
      enum: [raw, gst]
    BillLineItem:
      type: object
      properties:
        description:
          type: string
        qty:
          type: number
        rate:
          type: number
        amount:
          type: number
    Bill:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
        type:
          $ref: '#/components/schemas/BillType'
        seriesNumber:
          type: string
          example: TAX-2024-001
        amount:
          type: number
        isProtected:
          type: boolean
          description: Raw bills are password-protected
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/BillLineItem'
        taxableAmount:
          type: number
          description: GST bills only
        cgst:
          type: number
          description: GST bills only — 9%
        sgst:
          type: number
          description: GST bills only — 9%
        totalAmount:
          type: number
          description: GST bills only — taxableAmount + cgst + sgst
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
...
    BillList:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: '#/components/schemas/Bill'
        total:
          type: integer

    # Inventory ───────────────────────────────────────────────────────────────
...
    BarcodeType:
      type: string
      enum: [initial, final]
    Barcode:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
        type:
          $ref: '#/components/schemas/BarcodeType'
        code:
          type: string
        qrDataUrl:
          type: string
          description: Base64-encoded PNG data URL
        createdAt:
          type: string
          format: date-time
...
    BarcodeList:
      type: object
      properties:
        barcodes:
          type: array
          items:
            $ref: '#/components/schemas/Barcode'
        total:
          type: integer

    # Analytics ───────────────────────────────────────────────────────────────
...
    StatusCount:
      type: object
      properties:
        _id:
          type: string
        count:
          type: integer
    OrderStats:
      type: object
      properties:
        byStatus:
          type: array
          items:
            $ref: '#/components/schemas/StatusCount'
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      description: Requires `users.read`. Filter by role, availability, or active status.
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
        - name: isAvailable
          in: query
          schema:
            type: boolean
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Create a new user / staff account
      description: Requires `users.create`. Password is hashed server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Users]
      summary: Update user details
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Soft-delete a user (sets isActive=false)
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/availability:
    patch:
      tags: [Users]
      summary: Toggle staff availability (isAvailable)
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Orders ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      description: |
        Supports filtering by status, jobType, priority, and date range.
        Staff roles with `own` scope only see their own orders.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: jobType
          in: query
          schema:
            $ref: '#/components/schemas/JobType'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: from
          in: query
          description: ISO date lower bound (createdAt ≥ from)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: ISO date upper bound (createdAt ≤ to)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Orders]
      summary: Create a new order (starts in `draft`)
      description: |
        On creation the order number is auto-generated (`ORD-YYYY-NNNN`).
        Status is always `draft` — use `PATCH /{id}/status` to advance it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customer:
                name: Alice Corp
                phone: "9876543210"
                email: alice@corp.com
              jobType: flex_printing
              items:
                - description: Flex Banner 4x8
                  quantity: 2
                  unit: sqft
                  unitPrice: 250
              rawCost: 500
              taxableValue: 300
              billSplitPct: 40
              priority: normal
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Orders]
      summary: Update order fields (customer, costs, priority, etc.)
      description: Does not change status. Use `/status` for that.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Orders]
      summary: Cancel an order
      description: Transitions to `cancelled`. Triggers inventory reversal if stock was deducted.
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Advance the order through the state machine
      description: |
        **Valid transitions:**
        ```
        draft → confirmed → designing → in_production → finishing → completed → invoiced
        draft/confirmed/designing → cancelled
        ```
        **Side effects triggered automatically:**
        - `confirmed`: task is auto-assigned to the least-loaded available staff
        - `completed`: inventory deducted from BOM; final QR barcode generated
        - `cancelled`: previously deducted inventory is reversed
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetOrderStatusRequest'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid or disallowed transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid transition: draft → invoiced"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/timeline:
    get:
      tags: [Orders]
      summary: Get the full status history for an order
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Status history array
          content:
            application/json:
              schema:
                type: object
                properties:
                  timeline:
                    type: array
                    items:
                      $ref: '#/components/schemas/StatusHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Tasks ───────────────────────────────────────────────────────────────────
...
  /payments/order/{orderId}:
    get:
      tags: [Payments]
      summary: Get all payments for an order
      parameters:
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Payments for the order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /billing/order/{orderId}:
    get:
      tags: [Billing]
      summary: Get all bills for an order
      parameters:
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Bill list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /barcodes/order/{orderId}:
    get:
      tags: [Barcodes]
      summary: Get all barcodes for an order
      parameters:
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Barcodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BarcodeList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /analytics/orders:
    get:
      tags: [Analytics]
      summary: Order count grouped by status
      responses:
        '200':
          description: Order stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStats'
              example:
                byStatus:
                  - _id: draft
                    count: 12
                  - _id: confirmed
                    count: 8
                  - _id: completed
                    count: 34
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...